chore: move tasks to separate folder
