+++ apps/cli/src/tasks/diff.ts
@@ -1,0 +1,43 @@
+import * as ADCSDK from '@api7/adc-sdk';
+import { ListrTask } from 'listr2';
+
+import { TaskContext } from '../command/diff.command';
+import { DifferV3 } from '../differ/differv3';
+
+export const DiffResourceTask = (
+  printSummary = false,
+  persistentSummary = false,
+): ListrTask<TaskContext> => ({
+  title: 'Diff configuration',
+  task: async (ctx, task) => {
+    ctx.diff = DifferV3.diff(
+      ctx.local,
+      ctx.remote,
+      ctx.defaultValue,
+      undefined,
+      task,
+    );
+
+    if (printSummary) {
+      task.output = '';
+      let [created, updated, deleted] = [0, 0, 0];
+      ctx.diff.forEach((event) => {
+        switch (event.type) {
+          case ADCSDK.EventType.CREATE:
+            task.output += `create ${event.resourceType}: "${event.resourceName}"\n`;
+            created++;
+            break;
+          case ADCSDK.EventType.DELETE:
+            task.output += `delete ${event.resourceType}: "${event.resourceName}"\n`;
+            deleted++;
+            break;
+          case ADCSDK.EventType.UPDATE:
+            task.output += `update ${event.resourceType}: "${event.resourceName}"\n`;
+            updated++;
+            break;
+        }
+      });
+      task.output += `Summary: ${created} will be created, ${updated} will be updated, ${deleted} will be deleted`;
+    }
+  },
+});
