+++ apps/cli/src/tasks/index.ts
@@ -1,0 +1,5 @@
+export * from './load_local';
+export * from './load_remote';
+export * from './diff';
+export * from './lint';
+export * from './experimental';
